refactor(Systest): move systest initialization functions into SystestStarter and make SystestExecutor a class
