+++ nes-systests/systest/include/SystestExecutor.hpp
@@ -41,12 +41,16 @@ struct SystestExecutorResult
 namespace NES
 {
 
-void loadConfig(const argparse::ArgumentParser& program, SystestConfiguration& config);
-SystestConfiguration readConfiguration(int argc, const char** argv);
-void createSymlink(const std::filesystem::path& absoluteLogPath, const std::filesystem::path& symlinkPath);
-void setupLogging(const SystestConfiguration& config);
-void runEndlessMode(const std::vector<Systest::SystestQuery>& queries, SystestConfiguration& config);
+class SystestExecutor
+{
+public:
+    SystestExecutor(SystestConfiguration config) : config(config) { };
+    SystestExecutorResult executeSystests();
 
-SystestExecutorResult executeSystests(SystestConfiguration config);
+private:
+    void runEndlessMode(const std::vector<Systest::SystestQuery>& queries);
 
+    SystestConfiguration config;
+    Systest::SystestProgressTracker progressTracker;
+};
 }
