+++ nes-systests/systest/tests/SystestE2ETests.cpp
@@ -68,7 +68,8 @@ TEST_F(SystestE2ETest, CheckThatOnlyWrongQueriesFailInFileWithManyQueries)
     config.directlySpecifiedTestFiles.setValue(fmt::format("{}/errors/{}", SYSTEST_DATA_DIR, testFileName));
     config.workingDir.setValue(fmt::format("{}/nes-systests/systest/MultipleCorrectAndIncorrect", PATH_TO_BINARY_DIR));
 
-    const auto systestResult = NES::executeSystests(config);
+    SystestExecutor executor{config};
+    const auto systestResult = executor.executeSystests();
     ASSERT_TRUE(systestResult.returnType == SystestExecutorResult::ReturnType::FAILED) << " Return type not as expected.";
     ASSERT_FALSE(systestResult.outputMessage.contains(fmt::format("{}:1", testFileName))) << "Correct query found in failed queries.";
     ASSERT_TRUE(systestResult.outputMessage.contains(fmt::format("{}:2", testFileName))) << "Query not found in failed queries.";
@@ -94,7 +95,8 @@ TEST_P(SystestE2ETest, correctAndIncorrectSchemaTestFile)
     config.testFileExtension.setValue(std::string(EXTENSION));
     config.workingDir.setValue(fmt::format("{}/nes-systests/systest/{}", PATH_TO_BINARY_DIR, testFile));
 
-    const auto systestResult = NES::executeSystests(config);
+    SystestExecutor executor{config};
+    const auto systestResult = executor.executeSystests();
     ASSERT_TRUE(systestResult.returnType == SystestExecutorResult::ReturnType::FAILED) << " Return type not as expected.";
     ASSERT_EQ(countFailedTests(systestResult.outputMessage, SystestE2ETest::EXTENSION), 1) << "Too many failed queries.";
     ASSERT_FALSE(systestResult.outputMessage.contains(testFileName + std::string(":1"))) << "Correct query found in failed queries.";
